Speedup code-generation times
This speedup is achieved by removing the need for writing intermediary files just for the purpose of figuring out the dependencies that any given object has by means of inspecting said object's constructor arguments list.

With a different combination of reflection calls I was able to resolve this information directly from the user's script.